chore: move network-controller types
